Added question model to schema
