+++ prisma/schema.prisma
@@ -49,3 +49,12 @@ model users {
   password     String
   progress     Json
 }
+
+model questions {
+
+  id      String   @id @default(dbgenerated()) @map("_id") @db.ObjectId
+  tags    String[]
+  url     String?
+  options Json[]
+
+}
